refs #19; rename download module; removed useless options; added auto creating for temp directory; updated Makefile;
